Move AirlySensorEntityDescription to sensor platform (#55277)
